# docs/api.yml
openapi: 3.1.0
info:
  title: Laravel
  version: 0.0.1
servers:
  -
    url: 'http://localhost/api'
paths:
...
  /v1/notifications/unread-count:
    get:
      operationId: notification.unreadCount
      summary: 未読通知数取得
      tags:
        - Notification
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  unread_count: { type: integer }
                required:
                  - unread_count
        401:
          $ref: '#/components/responses/AuthenticationException'
  /v1/notifications/read-all:
    put:
      operationId: notification.markAllAsRead
      summary: すべての通知を既読にする
      tags:
        - Notification
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string, example: 'All notifications marked as read' }
                required:
                  - message
        401:
          $ref: '#/components/responses/AuthenticationException'
  '/v1/notifications/{notification}/read':
    put:
      operationId: notification.markAsRead
      summary: 通知を既読にする
      tags:
        - Notification
      parameters:
        -
          name: notification
          in: path
          required: true
          description: 'The notification ID'
          schema:
            type: integer
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  notification: { $ref: '#/components/schemas/NotificationResource' }
                required:
                  - notification
        403:
          description: 'Access denied'
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string, description: 'Error overview.' }
                required:
                  - message
        404:
          $ref: '#/components/responses/ModelNotFoundException'
        401:
          $ref: '#/components/responses/AuthenticationException'
...
components:
  schemas:
...
    NotificationResource:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        message:
          type:
            - string
            - 'null'
        type:
          type: string
        related_task_id:
          type:
            - integer
            - 'null'
        is_read:
          type: boolean
        read_at:
          type:
            - string
            - 'null'
          format: date-time
        created_at:
          type:
            - string
            - 'null'
          format: date-time
      required:
        - id
        - title
        - message
        - type
        - related_task_id
        - is_read
        - read_at
        - created_at
      title: NotificationResource
...
  responses:
...
    AuthenticationException:
      description: Unauthenticated
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: 'Error overview.'
            required:
              - message
    ModelNotFoundException:
      description: 'Not found'
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: 'Error overview.'
            required:
              - message